backend_but_better/pyproject.toml
[project]
name = "backend-but-better"
version = "0.1.0"
description = "FastAPI server bootstrap"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.116.0",
    "google-generativeai>=0.8.0",
    "lancedb>=0.22.0",
    "python-dotenv>=1.1.0",
    "uvicorn[standard]>=0.35.0",
]

[dependency-groups]
dev = [
    "httpx>=0.28.0",
    "pytest>=8.4.0",
]

[project.scripts]
start = "main:run"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.uv]
package = false
